remove clusterapi_template
