+++ platform-operator/controllers/verrazzano/component/clusterapi/clusterapi_overrides_test.go
@@ -167,3 +167,92 @@ func TestUserOverrides(t *testing.T) {
 	assert.Equal(t, "", oci.Version)
 	assert.Equal(t, "", oci.Url)
 }
+
+// TestTemplateInterface tests the OverridesInterface
+// GIVEN a set OverridesInput
+//
+//	WHEN the user supplies a OverridesInput
+//	THEN verify the OverridesInterface returns the expected values
+func TestTemplateInterface(t *testing.T) {
+	config.SetDefaultBomFilePath(testBomFilePath)
+
+	const capiOverrides = `
+{
+  "global": {
+    "registry": "ghcr.io"
+  },
+  "defaultProviders": {
+    "ocneBootstrap": {
+      "url": "/test/bootstrap.yaml",
+      "image": {
+        "registry": "myreg.io",
+        "tag": "v1.0"
+      }
+    },
+    "ocneControlPlane": {
+      "image": {
+        "tag": "v1.1"
+      }
+    },
+    "oci": {
+      "version": "v0.8.2",
+      "image": {
+        "repository": "repo",
+        "registry": "myreg2.io",
+        "tag": "v0.8.2"
+      }
+    },
+    "core": {
+    }
+  }
+}`
+
+	vz := &v1alpha1.Verrazzano{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "vz",
+		},
+		Spec: v1alpha1.VerrazzanoSpec{
+			Components: v1alpha1.ComponentSpec{
+				ClusterAPI: &v1alpha1.ClusterAPIComponent{
+					InstallOverrides: v1alpha1.InstallOverrides{
+						ValueOverrides: []v1alpha1.Overrides{
+							{
+								Values: &apiextensionsv1.JSON{
+									Raw: []byte(capiOverrides),
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	fakeClient := fake.NewClientBuilder().WithScheme(k8scheme.Scheme).WithObjects().Build()
+	compContext := spi.NewFakeContext(fakeClient, vz, nil, false)
+	config.TestHelmConfigDir = "../../../../helm_config"
+
+	templateInput, err := createTemplateInput(compContext)
+	assert.NoError(t, err)
+	assert.NotNil(t, templateInput)
+	tc := newTemplateContext(templateInput)
+
+	assert.Equal(t, "ghcr.io", tc.GetGlobalRegistry())
+
+	assert.Equal(t, "myreg.io/verrazzano", tc.GetOCNEBootstrapRepository())
+	assert.Equal(t, "v1.0", tc.GetOCNEBootstrapTag())
+	assert.Equal(t, "/test/bootstrap.yaml", tc.GetOCNEBootstrapURL())
+
+	assert.Equal(t, "ghcr.io/verrazzano", tc.GetOCNEControlPlaneRepository())
+	assert.Equal(t, "v1.1", tc.GetOCNEControlPlaneTag())
+	assert.Equal(t, "/verrazzano/capi/control-plane-ocne/v0.1.0/control-plane-components.yaml", tc.GetOCNEControlPlaneURL())
+
+	assert.Equal(t, "ghcr.io/verrazzano", tc.GetClusterAPIRepository())
+	assert.Equal(t, "v1.3.3-20230427222746-876fe3dc9", tc.GetClusterAPITag())
+	assert.Equal(t, "/verrazzano/capi/cluster-api/v1.3.3/core-components.yaml", tc.GetClusterAPIURL())
+
+	assert.Equal(t, "myreg2.io/repo", tc.GetOCIRepository())
+	assert.Equal(t, "v0.8.2", tc.GetOCITag())
+	assert.Equal(t, "https://github.com/verrazzano/capi/infrastructure-oci/v0.8.2/infrastructure-components.yaml", tc.GetOCIURL())
+
+}
